// src/models/alert.ts
export class Alert {

  id?: number;
  version?: number;
  alertType?: number;
  image?: number;
  message?: string;
  poubelle?: number;
  created?: string;
  favored?: boolean;
  checked?: boolean;
  deleted?: boolean;
  creator?: number;
  voted?: boolean;
  votes?: number;

  constructor() {}

}